Increase spice level
This change _arguably_ simplifies the record class, by removing the
need for a separate RecordStore to manage dynamic subclasses, and moving
that functionality into the metaclass.

Moreover, the switch to WeakValueDictionaries should eliminate a memory
leak if many different record types are created and then destroyed, as
their dynamic subclasses will now be destroyed along with them.